Statisitcs aggregators added, testcase updated

+++ notification_rest_service/src/main/java/net/bcnnm/notifications/slack/ReportsMean.java
@@ -27,9 +27,9 @@ public class ReportsMean implements ReportsAggregator{
         Method keyGetter = keyPropertyDescriptor.getReadMethod();
 
         double meanValue = reports.stream()
-                .mapToLong(report -> {
+                .mapToDouble(report -> {
                     try {
-                        return ((Number) keyGetter.invoke(report)).longValue();
+                        return ((Number) keyGetter.invoke(report)).doubleValue();
                     } catch (IllegalAccessException | InvocationTargetException | ClassCastException e) {
                         // LOG exception here
                         throw new AggregationException(String.format("Failed to get value of field: %s", key), e);
